server/requirements.txt
alembic
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
psycopg2-binary==2.9.10
email-validator
passlib[argon2]
python-jose[cryptography]
python-multipart
debugpy
